+++ server.js
@@ -123,6 +123,7 @@ app.use(bodyParser.json({limit: '50mb'}));
 
 //public
 app.use('/public',express.static('public'));
+app.use('/release',express.static('release'));
 app.use('/.well-known', express.static('.well-known'));
 
 //session control