+++ changelog.md
@@ -61,3 +61,23 @@ add:
 update:
     1. video widget
 fix:
+
+v1.8.8
+add:
+    1. remove duplicated files for rendering
+    2. pagination for blog
+    3. update for local ide
+update:
+    1. update blog category
+fix:
+
+v1.8.9
+add:
+    1. new IDE manage space
+    2. admin can release new local version in manage space
+    3. add download link to download local version
+    4. IDE support save project as XX in web
+update:
+fix:
+    1. fix a bug that versionTag didn't display in page
+    2. fix a bug that delete a project that mongodb cause server crash